add endpoint saftey to prevent server dying in production
