+++ server/utils/chroma/index.js
@@ -2,46 +2,20 @@ const { ChromaClient, OpenAIEmbeddingFunction } = require("chromadb");
 const { Chroma: ChromaStore } = require("langchain/vectorstores/chroma");
 const { OpenAI } = require("langchain/llms/openai");
 const { ChatOpenAI } = require("langchain/chat_models/openai");
-const {
-  VectorDBQAChain,
-  LLMChain,
-  RetrievalQAChain,
-  ConversationalRetrievalQAChain,
-} = require("langchain/chains");
+const { VectorDBQAChain } = require("langchain/chains");
 const { OpenAIEmbeddings } = require("langchain/embeddings/openai");
-// const { VectorStoreRetrieverMemory, BufferMemory } = require("langchain/memory");
-// const { PromptTemplate } = require("langchain/prompts");
 const { RecursiveCharacterTextSplitter } = require("langchain/text_splitter");
 const { storeVectorResult, cachedVectorInformation } = require("../files");
 const { Configuration, OpenAIApi } = require("openai");
 const { v4: uuidv4 } = require("uuid");
-
-const toChunks = (arr, size) => {
-  return Array.from({ length: Math.ceil(arr.length / size) }, (_v, i) =>
-    arr.slice(i * size, i * size + size)
-  );
-};
-
-function curateSources(sources = []) {
-  const knownDocs = [];
-  const documents = [];
-  for (const source of sources) {
-    const { metadata = {} } = source;
-    if (
-      Object.keys(metadata).length > 0 &&
-      !knownDocs.includes(metadata.title)
-    ) {
-      documents.push({ ...metadata });
-      knownDocs.push(metadata.title);
-    }
-  }
-
-  return documents;
-}
+const { toChunks, curateSources } = require("../helpers");
 
 const Chroma = {
-  name: 'Chroma',
+  name: "Chroma",
   connect: async function () {
+    if (process.env.VECTOR_DB !== "chroma")
+      throw new Error("Chroma::Invalid ENV settings");
+
     const client = new ChromaClient({
       path: process.env.CHROMA_ENDPOINT, // if not set will fallback to localhost:8000
     });
@@ -356,6 +330,4 @@ const Chroma = {
   },
 };
 
-module.exports = {
-  Chroma,
-};
+module.exports.Chroma = Chroma;
