+++ server/utils/helpers/index.js
@@ -1,7 +1,7 @@
-const { Pinecone } = require("../pinecone");
-const { Chroma } = require("../chroma");
-
 function getVectorDbClass() {
+  const { Pinecone } = require("../pinecone");
+  const { Chroma } = require("../chroma");
+
   const vectorSelection = process.env.VECTOR_DB || "pinecone";
   switch (vectorSelection) {
     case "pinecone":
@@ -9,10 +9,35 @@ function getVectorDbClass() {
     case "chroma":
       return Chroma;
     default:
-      return Pinecone;
+      throw new Error("ENV: No VECTOR_DB value found in environment!");
   }
+}
+
+function toChunks(arr, size) {
+  return Array.from({ length: Math.ceil(arr.length / size) }, (_v, i) =>
+    arr.slice(i * size, i * size + size)
+  );
+}
+
+function curateSources(sources = []) {
+  const knownDocs = [];
+  const documents = [];
+  for (const source of sources) {
+    const { metadata = {} } = source;
+    if (
+      Object.keys(metadata).length > 0 &&
+      !knownDocs.includes(metadata.title)
+    ) {
+      documents.push({ ...metadata });
+      knownDocs.push(metadata.title);
+    }
+  }
+
+  return documents;
 }
 
 module.exports = {
   getVectorDbClass,
+  toChunks,
+  curateSources,
 };
